turn blender host into module
